Redis server and offline token caching
Enable the server to dynamically switch between using the Redis server and its own offline in-memory cache for user tokens based on the availability of the Redis server. This is hidden behind an AuthService, where re-connect frequency is set. Adjust the API/endpoints to utilise the new AuthService instead of directly working with the Cache struct